relocate parser
