+++ src/lib/modules/response/resp_delay.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/zairza-cetb/bench-routes/src/lib/filters"
 	"github.com/zairza-cetb/bench-routes/src/lib/utils"
+	"github.com/zairza-cetb/bench-routes/src/lib/utils/parser"
 	"github.com/zairza-cetb/bench-routes/tsdb"
 )
 
@@ -20,13 +21,13 @@ const (
 // HandleResponseDelayForRoute is the initial entrypoint function for this module which takes
 // in a Route struct and supplies it to a function in turn to handle it accordingly. We create
 // channels to run tests for each route in parallel, speeding up the process
-func HandleResponseDelayForRoute(responseChains [][]*tsdb.Chain, route utils.Routes, tsdbNameHash string, wg *sync.WaitGroup) {
+func HandleResponseDelayForRoute(responseChains [][]*tsdb.Chain, route parser.Routes, tsdbNameHash string, wg *sync.WaitGroup) {
 	routeSuffix := filters.RouteDestroyer(route.URL)
 	// Init paths for request-response-monitoring
 	pathDelay := PathReqResDelay + "/" + "chunk_req_res_" + routeSuffix + "_delay.json"
 	pathLength := PathReqResDelay + "/" + "chunk_req_res_" + routeSuffix + "_length.json"
 	pathStatusCode := PathReqResDelay + "/" + "chunk_req_res_" + routeSuffix + "_status.json"
-	c := make(chan utils.Response)
+	c := make(chan parser.Response)
 	go RouteDispatcher(route, c)
 	responseObject := <-c
 	// Store the respective attributes of the
@@ -88,20 +89,20 @@ func HandleResponseDelayForRoute(responseChains [][]*tsdb.Chain, route utils.Rou
 }
 
 // RouteDispatcher dispatches a route to respective handlers based on it's request type
-func RouteDispatcher(route utils.Routes, c chan utils.Response) {
+func RouteDispatcher(route parser.Routes, c chan parser.Response) {
 	if route.Method == "GET" {
 		res := HandleGetRequest(route.URL)
 		c <- res
 	} else {
 		// Send a very large integer to automatically rule out as it
 		// is much much larger than the threshold
-		c <- utils.Response{Delay: math.MaxInt32, ResLength: 0, ResStatusCode: 100}
+		c <- parser.Response{Delay: math.MaxInt32, ResLength: 0, ResStatusCode: 100}
 	}
 }
 
 // HandleGetRequest specifically handles routes with GET Requests. Calculates timestamp before
 // and after processing of each request and returns the difference
-func HandleGetRequest(url string) utils.Response {
+func HandleGetRequest(url string) parser.Response {
 	// Time init
 	start := time.Now().UnixNano()
 	resp := utils.SendGETRequest(url)
@@ -112,5 +113,5 @@ func HandleGetRequest(url string) utils.Response {
 	end := time.Now().UnixNano()
 	diff := int((end - start) / int64(time.Millisecond))
 
-	return utils.Response{Delay: diff, ResLength: resLength, ResStatusCode: respStatusCode}
+	return parser.Response{Delay: diff, ResLength: resLength, ResStatusCode: respStatusCode}
 }
