--- storage/flood-ping/chunk_flood_ping_2033392131.json
@@ -1,1 +1,0 @@
-[]
