--- storage/flood-ping/chunk_flood_ping_3994026321.json
@@ -1,1 +1,0 @@
-[]
