+++ storage/jitter/chunk_jitter_google.co.in.json
@@ -1,1 +1,37 @@
-[{"datapoint":67,"normalizedTime":0,"timestamp":"2019-10-08T23:06:53.238487496+05:30"},{"datapoint":125,"normalizedTime":1,"timestamp":"2019-10-08T23:08:49.359568059+05:30"},{"datapoint":31.944445,"normalizedTime":2,"timestamp":"2019-10-10T18:00:31.208690931+05:30"},{"datapoint":157.77777,"normalizedTime":3,"timestamp":"2019-10-10T18:00:43.51987129+05:30"},{"datapoint":11.6,"normalizedTime":4,"timestamp":"2019-10-10T18:00:55.887218235+05:30"},{"datapoint":10.288889,"normalizedTime":5,"timestamp":"2019-10-10T18:01:08.242116987+05:30"},{"datapoint":3.5444446,"normalizedTime":6,"timestamp":"2019-10-10T18:01:20.566902304+05:30"}]
+[
+	{
+		"datapoint": 9.666667,
+		"normalizedTime": 0,
+		"timestamp": "2019-10-11T18:52:54.175425982+05:30"
+	},
+	{
+		"datapoint": 38.7,
+		"normalizedTime": 1,
+		"timestamp": "2019-10-11T18:53:06.444602276+05:30"
+	},
+	{
+		"datapoint": 11.855556,
+		"normalizedTime": 2,
+		"timestamp": "2019-10-11T18:53:18.640297658+05:30"
+	},
+	{
+		"datapoint": 7.3333335,
+		"normalizedTime": 3,
+		"timestamp": "2019-10-11T18:53:30.812799606+05:30"
+	},
+	{
+		"datapoint": 454.225,
+		"normalizedTime": 4,
+		"timestamp": "2019-10-11T18:53:45.578590128+05:30"
+	},
+	{
+		"datapoint": 36,
+		"normalizedTime": 5,
+		"timestamp": "2019-10-11T18:54:04.777283174+05:30"
+	},
+	{
+		"datapoint": 306.33334,
+		"normalizedTime": 6,
+		"timestamp": "2019-10-11T18:54:50.863289837+05:30"
+	}
+]
