--- storage/ping/chunk_ping_2828393398.json
@@ -1,1 +1,0 @@
-[]
