--- storage/ping/chunk_ping_3147873908.json
@@ -1,1 +1,0 @@
-[]
