--- storage/ping/chunk_ping_3994026321.json
@@ -1,1 +1,0 @@
-[]
