+++ storage/ping/chunk_ping_bing.com.json
@@ -1,62 +1,72 @@
 [
 	{
 		"datapoint": {
-			"Min": 67.641,
-			"Mean": 86.601,
-			"Max": 140.773,
-			"MDev": 19.5
+			"Min": 75.4,
+			"Mean": 86.866,
+			"Max": 96.298,
+			"MDev": 6.429
 		},
 		"normalizedTime": 0,
-		"timestamp": "2019-10-10T18:00:25.35147991+05:30"
+		"timestamp": "2019-10-11T18:52:48.486482262+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 65.124,
-			"Mean": 84.45,
-			"Max": 123.291,
-			"MDev": 15.222
+			"Min": 70.552,
+			"Mean": 85.376,
+			"Max": 99.414,
+			"MDev": 7.937
 		},
 		"normalizedTime": 1,
-		"timestamp": "2019-10-10T18:00:37.704619465+05:30"
+		"timestamp": "2019-10-11T18:53:01.88028785+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 69.102,
-			"Mean": 84.203,
-			"Max": 93.521,
-			"MDev": 7.608
+			"Min": 71.834,
+			"Mean": 83.19,
+			"Max": 100.725,
+			"MDev": 9.297
 		},
 		"normalizedTime": 2,
-		"timestamp": "2019-10-10T18:00:50.022355705+05:30"
+		"timestamp": "2019-10-11T18:53:14.176426957+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 62.788,
-			"Mean": 105.954,
-			"Max": 378.564,
-			"MDev": 91.013
+			"Min": 71.593,
+			"Mean": 79.828,
+			"Max": 92.912,
+			"MDev": 6.195
 		},
 		"normalizedTime": 3,
-		"timestamp": "2019-10-10T18:01:02.338947895+05:30"
+		"timestamp": "2019-10-11T18:53:26.419718887+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 68.98,
-			"Mean": 83.785,
-			"Max": 93.312,
-			"MDev": 8.348
+			"Min": 77.846,
+			"Mean": 96.963,
+			"Max": 201.376,
+			"MDev": 37.067
 		},
 		"normalizedTime": 4,
-		"timestamp": "2019-10-10T18:01:14.702662412+05:30"
+		"timestamp": "2019-10-11T18:53:39.62754513+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 67.223,
-			"Mean": 75.76,
-			"Max": 92.902,
-			"MDev": 7.795
+			"Min": 2926.577,
+			"Mean": 3054.143,
+			"Max": 3136.926,
+			"MDev": 77.97
 		},
 		"normalizedTime": 5,
-		"timestamp": "2019-10-10T18:01:27.019213087+05:30"
+		"timestamp": "2019-10-11T18:54:15.052780165+05:30"
+	},
+	{
+		"datapoint": {
+			"Min": 2400.903,
+			"Mean": 2731.631,
+			"Max": 3001.493,
+			"MDev": 215.176
+		},
+		"normalizedTime": 6,
+		"timestamp": "2019-10-11T18:54:47.038760157+05:30"
 	}
 ]
