+++ storage/ping/chunk_ping_facebook.com.json
@@ -1,62 +1,72 @@
 [
 	{
 		"datapoint": {
-			"Min": 74.003,
-			"Mean": 95.702,
-			"Max": 167.142,
-			"MDev": 27.121
+			"Min": 77.441,
+			"Mean": 86.244,
+			"Max": 98.042,
+			"MDev": 5.83
 		},
 		"normalizedTime": 0,
-		"timestamp": "2019-10-10T18:00:25.24858049+05:30"
+		"timestamp": "2019-10-11T18:52:48.269712116+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 70.472,
-			"Mean": 82.522,
-			"Max": 91.807,
-			"MDev": 6.533
+			"Min": 75.292,
+			"Mean": 80.096,
+			"Max": 86.211,
+			"MDev": 3.566
 		},
 		"normalizedTime": 1,
-		"timestamp": "2019-10-10T18:00:37.556654781+05:30"
+		"timestamp": "2019-10-11T18:53:01.70426508+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 69.708,
-			"Mean": 76.522,
-			"Max": 86.812,
-			"MDev": 6.296
+			"Min": 68.432,
+			"Mean": 79.398,
+			"Max": 94.878,
+			"MDev": 7.368
 		},
 		"normalizedTime": 2,
-		"timestamp": "2019-10-10T18:00:49.880082739+05:30"
+		"timestamp": "2019-10-11T18:53:14.014812661+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 70.417,
-			"Mean": 118.439,
-			"Max": 461.79,
-			"MDev": 114.546
+			"Min": 67.135,
+			"Mean": 76.24,
+			"Max": 83.611,
+			"MDev": 5.112
 		},
 		"normalizedTime": 3,
-		"timestamp": "2019-10-10T18:01:02.249456542+05:30"
+		"timestamp": "2019-10-11T18:53:26.27829912+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 69.831,
-			"Mean": 98.758,
-			"Max": 195.552,
-			"MDev": 33.794
+			"Min": 74.071,
+			"Mean": 109.48,
+			"Max": 218.238,
+			"MDev": 57.163
 		},
 		"normalizedTime": 4,
-		"timestamp": "2019-10-10T18:01:14.549034085+05:30"
+		"timestamp": "2019-10-11T18:53:39.449833994+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 69.753,
-			"Mean": 78.983,
-			"Max": 86.689,
-			"MDev": 4.791
+			"Min": 1599.367,
+			"Mean": 1625.758,
+			"Max": 1652.15,
+			"MDev": 26.422
 		},
 		"normalizedTime": 5,
-		"timestamp": "2019-10-10T18:01:26.897824588+05:30"
+		"timestamp": "2019-10-11T18:53:55.059887194+05:30"
+	},
+	{
+		"datapoint": {
+			"Min": 2432.361,
+			"Mean": 2868.555,
+			"Max": 3106.555,
+			"MDev": 235.436
+		},
+		"normalizedTime": 6,
+		"timestamp": "2019-10-11T18:54:33.429594582+05:30"
 	}
 ]
