+++ storage/ping/chunk_ping_google.co.in.json
@@ -1,62 +1,72 @@
 [
 	{
 		"datapoint": {
-			"Min": 85.066,
-			"Mean": 103.336,
-			"Max": 172.074,
-			"MDev": 25.249
+			"Min": 101.382,
+			"Mean": 106.03,
+			"Max": 112.121,
+			"MDev": 3.548
 		},
 		"normalizedTime": 0,
-		"timestamp": "2019-10-10T18:00:25.252901868+05:30"
+		"timestamp": "2019-10-11T18:52:48.341043419+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 77.308,
-			"Mean": 92.427,
-			"Max": 107.77,
-			"MDev": 9.845
+			"Min": 92.626,
+			"Mean": 100.146,
+			"Max": 110.099,
+			"MDev": 6.295
 		},
 		"normalizedTime": 1,
-		"timestamp": "2019-10-10T18:00:37.569576196+05:30"
+		"timestamp": "2019-10-11T18:53:01.709862741+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 83.787,
-			"Mean": 89.578,
-			"Max": 95.542,
-			"MDev": 3.729
+			"Min": 86.893,
+			"Mean": 94.471,
+			"Max": 103.382,
+			"MDev": 4.489
 		},
 		"normalizedTime": 2,
-		"timestamp": "2019-10-10T18:00:49.930210472+05:30"
+		"timestamp": "2019-10-11T18:53:14.030732621+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 87.129,
-			"Mean": 129.568,
-			"Max": 464.229,
-			"MDev": 111.65
+			"Min": 87.479,
+			"Mean": 95.135,
+			"Max": 105.025,
+			"MDev": 4.816
 		},
 		"normalizedTime": 3,
-		"timestamp": "2019-10-10T18:01:02.25949237+05:30"
+		"timestamp": "2019-10-11T18:53:26.311025838+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 77.172,
-			"Mean": 106.717,
-			"Max": 197.527,
-			"MDev": 32.023
+			"Min": 94.101,
+			"Mean": 128.857,
+			"Max": 246.167,
+			"MDev": 59.26
 		},
 		"normalizedTime": 4,
-		"timestamp": "2019-10-10T18:01:14.550872309+05:30"
+		"timestamp": "2019-10-11T18:53:39.451040589+05:30"
 	},
 	{
 		"datapoint": {
-			"Min": 75.791,
-			"Mean": 87.403,
-			"Max": 93.677,
-			"MDev": 4.9
+			"Min": 2645.775,
+			"Mean": 2865.167,
+			"Max": 3084.559,
+			"MDev": 219.392
 		},
 		"normalizedTime": 5,
-		"timestamp": "2019-10-10T18:01:26.897824589+05:30"
+		"timestamp": "2019-10-11T18:53:57.17500315+05:30"
+	},
+	{
+		"datapoint": {
+			"Min": 2463.187,
+			"Mean": 2903.341,
+			"Max": 3266.19,
+			"MDev": 246.413
+		},
+		"normalizedTime": 6,
+		"timestamp": "2019-10-11T18:54:33.697125655+05:30"
 	}
 ]
